Separate implementation to different files
